Ast normalizer (#2)
Add a normalizer that actually parses the sql query. output is a little different:

around whitespace
we drop the "asc" in order by-clauses
We also track the tables mentioned in the query, and also the toplevel operation (select, update, insert, delete, etc.)

if we fail to parse (which can happen with some valid mysql queries) we fall back to the scan normalizer.